+++ lib/constants.dart
@@ -1,8 +1,0 @@
-// BSD 3-Clause License
-// Copyright ©2022, GM Consult Pty Ltd
-// All rights reserved
-
-/// Exports the constants used in this package.
-library constants;
-
-export 'src/english_analyzer/english.dart' show EnglishConstants;