+++ lib/src/text_analyzer/_index.dart
@@ -2,6 +2,8 @@
 // Copyright ©2022, GM Consult Pty Ltd
 // All rights reserved
 
+export 'english_analyzer/english.dart';
+export 'latin_language_analyzer/latin_language_analyzer.dart';
 export 'term_co_occurrence_graph.dart';
 export 'text_analysis_extensions.dart';
 export 'text_analyzer.dart';