+++ lib/src/text_tokenizer/_index.dart
@@ -8,3 +8,4 @@ export 'n_gram_range.dart';
 // export 'text_document.dart';
 export 'text_tokenizer.dart';
 export 'token.dart';
+export 'tokenizer_strategy.dart';